+++ src/components/text/TextComponent.tsx
@@ -7,19 +7,26 @@ import { RichText } from "@kickstartds/base/lib/rich-text";
 export const Text: FC<TextProps & HTMLAttributes<HTMLElement>> = ({
   text,
   layout = "singleColumn",
+  align = "left",
   style = "default",
   ...props
 }) => {
-  const layoutClass = `${layout === "multiColumn" ? "c-text--columns" : ""}`;
-
-  const styleClass = `${style === "highlight" ? "c-text--highlight" : ""}`;
+  const layoutClass = `${
+    layout === "multiColumn" ? "c-rich-text--columns" : ""
+  }`;
+  const styleClass = `${style === "highlight" ? "c-rich-text--highlight" : ""}`;
+  const alignClass = `${align === "center" ? "c-rich-text--center" : ""}`;
 
   return (
-    <div
-      className={classnames("c-text", `${layoutClass}`, `${styleClass}`)}
+    <RichText
+      className={classnames(
+        "c-text",
+        `${layoutClass}`,
+        `${alignClass}`,
+        `${styleClass}`
+      )}
       {...props}
-    >
-      <RichText text={text} />
-    </div>
+      text={text}
+    />
   );
 };