+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# v2.0.16 (Wed Oct 25 2023)
+
+#### ⚠️ Pushed to `main`
+
+- feat(text): add alignment to text ([@fleven-kds](https://github.com/fleven-kds))
+
+#### Authors: 1
+
+- Franz ([@fleven-kds](https://github.com/fleven-kds))
+
+---
+
+# v2.0.15 (Wed Oct 25 2023)
+
+#### ⚠️ Pushed to `main`
+
+- feat(section): add new section styles ([@fleven-kds](https://github.com/fleven-kds))
+
+#### Authors: 1
+
+- Franz ([@fleven-kds](https://github.com/fleven-kds))
+
+---
+
 # v2.0.14 (Wed Oct 25 2023)
 
 #### ⚠️ Pushed to `main`